# Make GetEqualityStrategy thread-safe again