@jnyrup 2018-06-14
I've renamed the test to `When_concurrently_getting_equality_strategy_it_should_not_throw` and changed it to test directly on `GetEqualityStrategy()`.